+++ graphix/graphsim/__init__.py
@@ -1,0 +1,1 @@
+from .graphstate import GraphState
