+++ tests/test_graphsim.py
@@ -2,7 +2,7 @@ import unittest
 
 import numpy as np
 
-from graphix.graphsim import GraphState
+from graphix.graphsim.graphstate import GraphState
 from graphix.ops import Ops
 from graphix.sim.statevec import Statevec, meas_op
